major version

+++ src/include/HG.h
@@ -8,9 +8,10 @@
 #pragma once
 
 #include<pthread.h>
+#include "util.h"
 
 typedef struct {
-    size_t db_size;
+    integer db_size;
     char *db_hg_as0;
     char *db_hg_as1;
     char *db_hg_prev;
