+++ src/include/cou.h
@@ -7,13 +7,13 @@
 
 #pragma once
 
-#include "src/system/util.h"
+#include "util.h"
 
 typedef struct {
-    size_t db_size;
+    integer db_size;
     char *db_cou_primary;
     char *db_cou_shandow;
-    unsigned char*db_cou_curBA;
+    unsigned char *db_cou_curBA;
     unsigned char *db_cou_preBA;
     unsigned char *db_cou_chgBA;
     unsigned char *db_cou_access;
